LayoutTests/editing/spelling/spelling-with-whitespace-selection-expected.txt

Spelling should ignore whitespace. To test manually, type 'hello, wellcome home.', select ' wellcome ', and context-click on the selection. The test succeeds when the context menu has spell check suggestions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is " wellcome "
PASS spellingMarkerRange.toString() is "wellcome"
PASS window.getSelection().toString() is " wellcome "
PASS contextMenuElements[contextMenuElements.length - 1] is "welcome"
PASS successfullyParsed is true

TEST COMPLETE
hello, wellcome home.
